Set 4

+++ Problem_set_3/14.c
@@ -2,46 +2,46 @@
 #include<math.h>
 
 int main() {
-    float a, b;
+    double a, b;
     int choice, divisionChoice;
-    scanf("%f %f", &a, &b);
+    scanf("%lf %lf", &a, &b);
     scanf("%d", &choice);
 
-    if (choice == 1) 
+    if (choice == 1)
     {
-        printf("Addition: %.2f\n", a + b);
+        printf("Addition: %.2lf\n", a + b);
     }
-     else if (choice == 2) 
+     else if (choice == 2)
     {
-        printf("Subtraction: %.2f\n", a - b);
+        printf("Subtraction: %.2lf\n", a - b);
     }
-     else if (choice == 3) 
+     else if (choice == 3)
     {
-        printf("Multiplication: %.2f\n", a * b);
-    } 
-    else if (choice == 4) 
+        printf("Multiplication: %.2lf\n", a * b);
+    }
+    else if (choice == 4)
     {
         if (b != 0) {
-            
+
             scanf("%d", &divisionChoice);
-            
-            if (divisionChoice == 1) 
+
+            if (divisionChoice == 1)
             {
-                printf("Quotient: %.2f\n", a / b);
-            } 
-            else if (divisionChoice == 2) 
+                printf("Quotient: %.2lf\n", a / b);
+            }
+            else if (divisionChoice == 2)
             {
-                printf("Reminder: %.2f\n", fmod(a, b));
-            } 
-            else 
+                printf("Reminder: %.2lf\n", fmod(a, b));
+            }
+            else
             {
                 printf("Invalid division choice! Please enter a valid choice.\n");
             }
-        } 
+        }
         else {
             printf("Error: Divisor is zero\n");
         }
-    } 
+    }
     else {
         printf("Invalid choice! Please enter a valid choice.\n");
     }
